Migrated timer to typescript
